+++ tools/scripts/get-ci-image-tag.sh
@@ -2,6 +2,7 @@
 set -euo pipefail
 
 # Source shared utility functions
+# shellcheck source=tools/scripts/functions.sh
 source "$(dirname "${BASH_SOURCE[0]}")/functions.sh"
 
 # Accepts env vars already loaded in the calling context
@@ -9,11 +10,9 @@ main() {
   local container_hash
   container_hash=$(get_container_hash)
 
-  local tag_vars="RUST_$RUST_VERSION-ZCASH_$ZCASH_VERSION-ZEBRA_$ZEBRA_VERSION-CONTAINER_$container_hash"
-  local tag
-  tag=$(echo "$tag_vars" | tr ' ' '\n' | sort | sha256sum | cut -c1-12)
-  # echo "VERSIONS: $tag_vars"
-  # echo "TAG: $tag"
+  local tag_vars
+  tag_vars="RUST_$RUST_VERSION-ZCASH_$ZCASH_VERSION"
+  tag_vars+="-ZEBRA_$ZEBRA_VERSION-CONTAINER_$container_hash"
   echo "$tag_vars"
 }